Complete exercise 04
This was a tougher one. Avatar and proportioned / distorted classes
had to be chained together. Because we were changing the ones on
top, we needed to select the avatar class, not original. That is
the mistake I was making. From there, the auto vs. hard-coded px
size is what makes the difference.