Complete flex ex04
This was solved by making flexoxes within flexboxes. Made and items
class to have all the items flex horizontal but then an items class
to wrap each individual item in. The tricky part was remembering how
the axis changes when you flip to column, so I had to use align-items
instead of justify-content in order to center the text and images.